telecom_ops/requirements.txt
google-adk==1.14.0
toolbox-core
sqlalchemy
pg8000
